First commit.
Separation of CalcIncl. This tool only includes the caluclation of masks from PLI measurements.
The performance of the mask calculation should be similar or better as previous limitations are still in tact. Those should be adressed in future versions.

TODO: LondonRat s0126 and s0121 result in erronous mask due to the retardation. Normally only one peak is present. A second peak can also show up when the background and gray matter show values that are vastly different from each other. However, in those measurements a plateau is visible which results in too low values for t_ret. This error should be adressed next

Signed-off-by: Jan André Reuter <jan.andre.reuter@hotmail.de>