Added tests, fixed bugs.
Fixed a bug in mask.py which could result in too high values for t_ret. Those values did seem good in the results which should be looked at when comparing the new results. For now, this code will pass the provided tests.
Fixed a bug in test_cmd.py where the wrong paths were used. Therefore not all paths were deleted after all tests were executed.
Added some of the tests in mask.py. This will be expanded in the coming versions.

Signed-off-by: Jan André Reuter <jan.andre.reuter@hotmail.de>