Fixed a few bugs with the inclination calculation.
The inclination will now be generated as expected when calculating more than one inclination at a time. The background will be set to 90° as in the Python version.
Changed PLImg.cpp to use the same parameter layout as CalcIncl.cpp (itra, iret, ...) which allows that the base files are in different folders. This allows to use our database setup without the need of soft links or a duplicate in a separate folder.

Signed-off-by: Jan André Reuter <jan.andre.reuter@hotmail.de>