Fix for tRet() calculation
Previously tRet() was heavily influenced by plateaus and nearby peaks resulting in faulty results. The last few commits all tried to fix this issue by changing the calculation formula and trying different approaches to define the maximum curvature.

This commit should fix the erronous results by calculating the histogram for many different bin sizes (16 - 256 bins) and calculating tRet() for each bin size. Starting at 16 bins tRet() will be calculated normally. The results will be used as an estimate in higher bin counts. In each following iterating the bin count will be doubled and a region nearby the estimation will only be considered as the maximum curvature.

Tests will still need to be fixed. Also, the blurred mask seems to result in "not so great" values which has to be verified through the inclination. This will be looked at next.

Signed-off-by: Jan André Reuter <jan.andre.reuter@hotmail.de>