Fix for some measurements
This commit should fix another round of problems which are happening because of weird OpenCV parameter choices. This results in "correct" parameter maps for PE-2012-00102-V_00_s0626 to PE-2012-00102_V_00_s0636 and PE-2018-00516-R_00_s0110 to PE-2018-00516_00_s129 (at least for the Retardation).

The next commits will check if the transmittance parameters are reliable.

Signed-off-by: Jan André Reuter <jan.andre.reuter@hotmail.de>