Project renamed. Added evaluation of blurred mask.
The Python script generates two files per parameter of our blurred mask
which represent the mean value and standard deviation in a boxplot as
well as normal Python plot. The tool does not use any kind of argument
parsing yet. Therefore the result files are written to /tmp while the
first parameter is used as the input folder.

Signed-off-by: Jan Andre Reuter <j.reuter@fz-juelich.de>