Changed median transmittance test (still failing..)
Fixed a bug which could cause a crash of the program because the actually needed memory wasn't calculated correctly.

The crash actually happened becuase NVIDIA NPP does return 'int' values for their buffer sizes which is not enough when dealing with large images. The buffer size is now calculated without their buffer methods as the approximately needed memory can be caluclated manually.

Signed-off-by: Jan André Reuter <jan.andre.reuter@hotmail.de>